Time: 4 ms (12.46%) | Memory: 29.9 MB (8.78%) - LeetSync

+++ 3251-maximum-area-of-longest-diagonal-rectangle/maximum-area-of-longest-diagonal-rectangle.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int areaOfMaxDiagonal(vector<vector<int>>& dimensions) {
-        double diagonal = INT_MIN;
+        float diagonal = INT_MIN;
         int area = INT_MIN;
-        double maxi = INT_MIN;
+        float maxi = INT_MIN;
         for(auto it : dimensions){
             diagonal = sqrt(it[0]*it[0] + it[1]*it[1]);
             if(diagonal > maxi) {
